narwhalified cappingtransformer - moved tests to base capping transformer file, where they fit better
