+++ tests/capping/test_BaseCappingTransformer.py
@@ -350,6 +350,52 @@ class GenericCappingFitTests(WeightColumnFitMixinTests, GenericFitTests):
                 actuals_dict[name] == value
             ), f"unexpected replacement values fit, for {name} value expected {value} but got {actuals_dict[name]}"
 
+    @pytest.mark.parametrize("library", ["pandas", "polars"])
+    @pytest.mark.parametrize(
+        ("values", "sample_weight", "quantiles"),
+        [
+            ([1, 2, 3], [1, 2, 1], [0.1, 0.5]),
+            ([1, 2, 3], None, [0.1, 0.5]),
+            ([2, 4, 6, 8], [3, 2, 1, 1], [None, 0.5]),
+            ([2, 4, 6, 8], None, [None, 0.5]),
+            ([-1, -5, -10, 20, 30], [1, 2, 2, 2, 2], [0.1, None]),
+            ([-1, -5, -10, 20, 40], None, [0.1, None]),
+        ],
+    )
+    def test_replacement_values_updated(
+        self,
+        values,
+        sample_weight,
+        quantiles,
+        minimal_attribute_dict,
+        uninitialized_transformers,
+        library,
+    ):
+        """Test that weighted_quantile gives the expected outputs."""
+
+        args = minimal_attribute_dict[self.transformer_name].copy()
+        args["quantiles"] = {"a": quantiles}
+        args["capping_values"] = None
+        args["weights_column"] = "w"
+
+        transformer = uninitialized_transformers[self.transformer_name](**args)
+
+        if not sample_weight:
+            sample_weight = [1] * len(values)
+
+        df_dict = {
+            "a": values,
+            "w": sample_weight,
+        }
+
+        df = dataframe_init_dispatch(dataframe_dict=df_dict, library=library)
+
+        transformer.fit(df)
+
+        assert (
+            transformer._replacement_values == transformer.quantile_capping_values
+        ), f"unexpected value for replacement_values attribute, expected {transformer.quantile_capping_values} but got {transformer.replacement_values_}"
+
 
 class GenericCappingTransformTests(GenericTransformTests):
     """Tests for BaseCappingTransformer.transform()."""
@@ -676,6 +722,49 @@ class GenericCappingTransformTests(GenericTransformTests):
         assert (
             transformer.quantiles == transformer2.quantiles
         ), "quantiles attribute modified in transform"
+
+    def expected_df_1(self, library="pandas"):
+        """Expected output from test_expected_output_min_and_max."""
+        df_dict = {
+            "a": [2, 2, 3, 4, 5, 5, None],
+            "b": [1, 2, 3, None, 7, 7, 7],
+            "c": [None, 1, 2, 3, 0, 0, 0],
+        }
+
+        return dataframe_init_dispatch(dataframe_dict=df_dict, library=library)
+
+    @pytest.mark.parametrize("library", ["pandas", "polars"])
+    def test_expected_output_min_and_max_combinations(
+        self,
+        minimal_attribute_dict,
+        uninitialized_transformers,
+        library,
+    ):
+        """Test that capping is applied correctly in transform."""
+
+        df = d.create_df_3(library=library)
+        expected = self.expected_df_1(library=library)
+
+        args = minimal_attribute_dict[self.transformer_name].copy()
+        args["capping_values"] = {"a": [2, 5], "b": [None, 7], "c": [0, None]}
+
+        transformer = uninitialized_transformers[self.transformer_name](**args)
+
+        df_transformed = transformer.transform(df)
+
+        assert_frame_equal_dispatch(df_transformed, expected)
+
+        # Check outcomes for single rows
+        df = nw.from_native(df)
+        expected = nw.from_native(expected)
+        for i in range(len(df)):
+            df_transformed_row = transformer.transform(df[[i]].to_native())
+            df_expected_row = expected[[i]].to_native()
+
+            assert_frame_equal_dispatch(
+                df_transformed_row,
+                df_expected_row,
+            )
 
 
 class TestBaseCappingTransformerInit(GenericCappingInitTests):
